+++ src/Components/Header.css
@@ -8,14 +8,11 @@
     flex-direction: row;
     background: #000080;
     color: #ffffff;
-    height: 150px;
+    height: 100px;
   }
   
   .page_name{
-    font-size: 2.5em;
-  }
-
-  .date {
+    font-size: 3em;
     display: flex;
     flex-direction: column;
     flex-wrap: nowrap;
@@ -24,17 +21,16 @@
     margin-bottom: 25px;
     margin-top: 25px;
     margin-right: 2%;
-    margin-left: 83%;
-    font-size: 1.25em;
+    margin-left: 80%;
     font-weight: bold;
   }
 
   .back_button{
     position: absolute;
-    top: 37.5px;
+    top: 25px;
     left: 15px;
-    width: 75px;
-    height: 75px;
+    width: 50px;
+    height: 50px;
     border-radius: 20%;
     background: #ffffff;
     cursor: pointer;
@@ -42,7 +38,6 @@
   }
 
   .backimage img{
-    height: 50px;
-    width: 50px;
+    height: 35px;
+    width: 35px;
   }
-
